refactor: extract common types to shared package
