+++ packages/asset-utils/src/types.ts
@@ -1,31 +1,4 @@
-// TODO: app에서 선언된 타입이랑 중복되는거 같으니 common 패키지로 추출하기
-
-import { BlockStateApplyModelInfo } from '@depl/shared'
-
-export type BlockstatesFile =
-  | {
-      variants:
-        | Record<string, BlockStateApplyModelInfo>
-        | Record<string, BlockStateApplyModelInfo[]>
-    }
-  | {
-      multipart: {
-        apply: BlockStateApplyModelInfo | BlockStateApplyModelInfo[]
-        when?:
-          | {
-              AND: Record<string, string>[]
-            }
-          | {
-              OR: Record<string, string>[]
-            }
-          | Record<string, string>
-      }[]
-    }
-
-export interface ModelFile {
-  parent?: string
-  elements?: unknown[]
-}
+export type { BlockStatesFile, ModelFile } from '@depl/shared'
 
 // CDN files
