+++ packages/backend/src/types.ts
@@ -1,3 +1,8 @@
+export type {
+  TextureValue,
+  BackendAPIV1GetPlayerSkinResponse,
+} from '@depl/shared'
+
 export interface ProfileLookupData {
   id: string // player uuid
   name: string // player username
@@ -13,33 +18,4 @@ export interface PlayerSkinQueryResult {
   }[]
 }
 
-// base64 decoded value from PlayerSkinQueryResult
-export interface TextureValue {
-  timestamp: number
-  profileId: string // player uuid
-  profileName: string // player username
-  signatureRequired: boolean | undefined
-  textures: {
-    SKIN:
-      | {
-          url: string
-          metadata:
-            | {
-                model: 'slim' | undefined // wide skin = undefined
-              }
-            | undefined
-        }
-      | undefined
-    CAPE:
-      | {
-          url: string
-        }
-      | undefined
-  }
-}
-
-export interface APIGetPlayerSkinResponse {
-  id: string // player uuid
-  name: string // player username
-  skinUrl: string | null
-}
+// TextureValue = base64 decoded value from PlayerSkinQueryResult
